# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "SuperMakers"
        }
    ],
    "paths": {
...
        "/api/videos": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "Get a list of videos",
                "description": "Returns a list of all videos.",
                "operationId": "2090c29a8b2a49778e2de02f54ec7606",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Videos"
                ],
                "summary": "Create a new video",
                "description": "Adds a new video to the database",
                "operationId": "904069e3a71065b7227cc44e7dc50fb2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "url"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Sample Video"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "This is a sample video description."
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "http://example.com/sample-video.mp4"
                                    },
                                    "thumbnail": {
                                        "description": "Thumbnail image file upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Video created"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/videos/{id}": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "Get a video by ID",
                "description": "Returns a single video",
                "operationId": "d5190024e38b00df6680e59ca8a482b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the video to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video found"
                    },
                    "404": {
                        "description": "Video not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Videos"
                ],
                "summary": "Update an existing video",
                "description": "Updates video details by ID",
                "operationId": "296d69c432834f8ae560cc6917ad3d87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the video to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Video Title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "This is an updated video description."
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "http://example.com/updated-video.mp4"
                                    },
                                    "thumbnail": {
                                        "description": "Updated thumbnail file upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Video updated"
                    },
                    "404": {
                        "description": "Video not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Videos"
                ],
                "summary": "Delete a video",
                "description": "Deletes a video by ID",
                "operationId": "b8887a2f4df0b20670119bbb04b51e11",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the video to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Video deleted"
                    },
                    "404": {
                        "description": "Video not found"
                    }
                }
            }
        }
    },
...
}